+++ jest.config.js
@@ -10,5 +10,5 @@ module.exports = {
   coverageDirectory: "coverage",
 
   // The glob patterns Jest uses to detect test files
-  testMatch: ["**/*.test.ts"],
+  testMatch: ["**/__tests__/*.test.ts"],
 };