+++ src/AppRouter.tsx
@@ -1,4 +1,4 @@
-import { NavigationContainer, NavigatorScreenParams } from "@react-navigation/native";
+import { NavigatorScreenParams } from "@react-navigation/native";
 import {
   createNativeStackNavigator,
   NativeStackNavigationProp,
@@ -7,10 +7,16 @@ import React, { ReactElement } from "react";
 
 import { HomeScreen } from "@screens/Home";
 import { SettingsRouter } from "@screens/Settings";
-import { SettingsRouterParams } from "@screens/Settings/SettingsRouter";
+import { type SettingsRouterParams } from "@screens/Settings/SettingsRouter";
+import { type Day } from "@typings/day.types";
+
+import { useAppOpenLink } from "./app-link.hook";
 
 export type RootRouterParams = {
-  HomeScreen: undefined;
+  HomeScreen: {
+    /** Shared day parsed from external link  */
+    sharedDay?: Day;
+  };
   SettingsRouter: NavigatorScreenParams<SettingsRouterParams> | undefined;
 };
 
@@ -22,13 +28,14 @@ const AppRouter = (): ReactElement => {
   // TODO: Investigate adapting React Navigation theme to use MD3 colors
   // Source: https://callstack.github.io/react-native-paper/theming.html
 
+  useAppOpenLink();
+
+  // NOTE: Main 'NavigationContainer' is rendered by parent to allow accessing router here
   return (
-    <NavigationContainer>
-      <Stack.Navigator initialRouteName="HomeScreen" screenOptions={{ headerShown: false }}>
-        <Stack.Screen component={HomeScreen} name="HomeScreen" />
-        <Stack.Screen component={SettingsRouter} name="SettingsRouter" />
-      </Stack.Navigator>
-    </NavigationContainer>
+    <Stack.Navigator initialRouteName="HomeScreen" screenOptions={{ headerShown: false }}>
+      <Stack.Screen component={HomeScreen} name="HomeScreen" />
+      <Stack.Screen component={SettingsRouter} name="SettingsRouter" />
+    </Stack.Navigator>
   );
 };