+++ src/components/days/DayListItem.tsx
@@ -1,10 +1,11 @@
 import { MaterialCommunityIcons as Icon } from "@expo/vector-icons";
 import dayjs from "dayjs";
-import React, { ReactElement } from "react";
+import React, { Fragment, ReactElement } from "react";
 import { useTranslation } from "react-i18next";
 import { StyleProp, StyleSheet, View, ViewStyle } from "react-native";
-import { Avatar, Surface, Text, TouchableRipple, useTheme } from "react-native-paper";
+import { Avatar, Surface, Text, TouchableRipple } from "react-native-paper";
 
+import { useAppTheme } from "@hooks";
 import { sharedColors } from "@styles/theme";
 import { getDayDisplay } from "@utilities/day.util";
 
@@ -19,7 +20,7 @@ type DayDisplayProps = {
 const DayListItem = (props: DayDisplayProps): ReactElement | null => {
   const { day, style, onLongPress } = props;
 
-  const { colors } = useTheme();
+  const { colors } = useAppTheme();
   const { t } = useTranslation(["common", "screens"]);
 
   const dateDisplay = t("screens:home.listItemDate", {
@@ -46,7 +47,15 @@ const DayListItem = (props: DayDisplayProps): ReactElement | null => {
     >
       <Surface
         elevation={1}
-        style={[style, styles.day, { backgroundColor: backgroundColor, borderColor: mainColor }]}
+        style={[
+          style,
+          styles.day,
+          {
+            backgroundColor: backgroundColor,
+            borderColor: mainColor,
+            borderWidth: isToday ? 4 : 2,
+          },
+        ]}
       >
         <Avatar.Icon
           color={backgroundColor}
@@ -70,26 +79,40 @@ const DayListItem = (props: DayDisplayProps): ReactElement | null => {
         </View>
 
         <View style={[styles.dayStats, { backgroundColor: mainColor }]}>
-          {!isToday && (
-            <View
-              style={[
-                styles.dayStatsIconOutline,
-                { backgroundColor: countingDown ? mainColor : backgroundColor },
-              ]}
-            >
+          {!isToday ? (
+            <Fragment>
+              <View
+                style={[
+                  styles.dayStatsIconOutline,
+                  { backgroundColor: countingDown ? mainColor : backgroundColor },
+                ]}
+              >
+                <Icon
+                  color={countingDown ? backgroundColor : mainColor}
+                  name={countingDown ? "menu-down" : "menu-up"}
+                  style={styles.dayStatsIcon}
+                />
+              </View>
+              <Text
+                style={[styles.dayStatsCount, { color: mainColorText }]}
+                variant="headlineMedium"
+              >
+                {Math.abs(dateCount.count)}
+              </Text>
+              <Text style={[styles.dayStatsUnit, { color: mainColorText }]} variant="bodySmall">
+                {t("common:timeUnits.days", { count: Math.abs(dateCount.count) })}
+              </Text>
+            </Fragment>
+          ) : (
+            <Fragment>
               <Icon
-                color={countingDown ? backgroundColor : mainColor}
-                name={countingDown ? "menu-down" : "menu-up"}
-                style={styles.dayStatsIcon}
+                color={colors.warning}
+                // name="calendar-check"
+                name="alert-decagram"
+                style={styles.dayStatsTodayIcon}
               />
-            </View>
+            </Fragment>
           )}
-          <Text style={[styles.dayStatsCount, { color: mainColorText }]} variant="headlineMedium">
-            {Math.abs(dateCount.count)}
-          </Text>
-          <Text style={[styles.dayStatsUnit, { color: mainColorText }]} variant="bodySmall">
-            {t("common:timeUnits.days", { count: Math.abs(dateCount.count) })}
-          </Text>
         </View>
       </Surface>
     </TouchableRipple>
@@ -105,7 +128,6 @@ const styles = StyleSheet.create({
     flexDirection: "row",
     borderRadius: 8,
     overflow: "hidden",
-    borderWidth: 2,
   },
   dayContent: {
     flexGrow: 1,
@@ -150,7 +172,11 @@ const styles = StyleSheet.create({
     left: -(iconFontSize / 2 - iconFontOffset),
     borderRadius: iconFontSize * 2,
   },
+  dayStatsTodayIcon: {
+    fontSize: iconFontSize * 1.25,
+  },
   dayStatsUnit: {
+    // TODO: Determine whether unit should be centered with day count (offsets it) or not
     position: "absolute",
     bottom: 4,
     opacity: 0.8,