+++ src/errors/base.error.ts
@@ -1,0 +1,7 @@
+/** Base error class for internal exceptions */
+export abstract class BaseError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = this.constructor.name;
+  }
+}