+++ src/errors/index.ts
@@ -1,0 +1,2 @@
+export { BaseError } from "./base.error";
+export { CodedError } from "./coded.error";