+++ src/screens/Home/HomeScreen.tsx
@@ -1,24 +1,36 @@
-import { useNavigation } from "@react-navigation/native";
+import { useNavigation, useRoute } from "@react-navigation/native";
+import { NativeStackScreenProps } from "@react-navigation/native-stack";
 import dayjs from "dayjs";
-import React, { ReactElement, useRef, useState } from "react";
+import * as Linking from "expo-linking";
+import React, { ReactElement, useCallback, useEffect, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { StyleSheet, Vibration, View } from "react-native";
+import { Share, StyleSheet, Vibration, View } from "react-native";
 import { Text } from "react-native-paper";
 
-import { BottomSheetRef, DeleteDayDialog, ManageDaySheet } from "@components/dialogs";
+import {
+  AddSharedDayDialog,
+  BottomSheetRef,
+  DeleteDayDialog,
+  ManageDaySheet,
+} from "@components/dialogs";
 import { AppBar, Page, ScreenFAB } from "@components/layout";
 import { useAppDispatch, useAppSelector, useAppTheme, useScrollingFab, useSnackbar } from "@hooks";
 import { addDay, moveDay, removeDay, selectDays, updateDay } from "@store/slices/days";
+import { selectBehaviours } from "@store/slices/settings";
+import { SHARED_DAY_VERSION } from "@utilities/day-parse.util";
 
 import DayList from "./DayList";
 import SelectedDayModal from "./SelectedDayModal";
 
 import type { UpDown } from "@typings/app.types";
 import type { Day, DayNew } from "@typings/day.types";
-import type { RootRouterNavigation } from "src/AppRouter";
+import type { RootRouterParams } from "src/AppRouter";
+
+type HomeScreenRouteProps = NativeStackScreenProps<RootRouterParams, "HomeScreen">;
 
 const HomeScreen = (): ReactElement | null => {
-  const navigation = useNavigation<RootRouterNavigation>();
+  const navigation = useNavigation<HomeScreenRouteProps["navigation"]>();
+  const route = useRoute<HomeScreenRouteProps["route"]>();
 
   const { t } = useTranslation(["common", "screens"]);
   const { notify } = useSnackbar();
@@ -31,7 +43,7 @@ const HomeScreen = (): ReactElement | null => {
   const selectedDayPosition = selectedDay
     ? { count: days.length, position: days.findIndex((d) => d.id === selectedDay.id) + 1 }
     : null;
-  const dayOptionsRef = useRef<BottomSheetRef>(null);
+  const selectedDayRef = useRef<BottomSheetRef>(null);
 
   // FAB should disappear when scrolling down and reappear when scrolling back up
   const { fabVisible, toggleFab, onListScroll } = useScrollingFab();
@@ -43,13 +55,42 @@ const HomeScreen = (): ReactElement | null => {
 
   const today = dayjs();
 
+  const appBehaviours = useAppSelector(selectBehaviours);
+  const [sharedDay, setSharedDay] = useState<Day | null>(null);
+
+  const onSharedDayConfirm = useCallback(
+    (day: Day) => {
+      dispatch(addDay(day));
+      notify(t("screens:daySharedLink.dayAddSuccess", { title: day.title }));
+      setSharedDay(null);
+    },
+    [dispatch, notify, t],
+  );
+
+  useEffect(() => {
+    const day = route.params?.sharedDay;
+    if (!day) return;
+
+    if (!appBehaviours.confirmSharedDays) {
+      onSharedDayConfirm(day);
+    } else {
+      // Must wait briefly for navigation to occur before displaying dialog
+      setTimeout(() => {
+        setSharedDay(day);
+      }, 100);
+    }
+
+    // Avoid re-handling shared day again on route re-render (if param is kept)!
+    navigation.setParams({ sharedDay: undefined });
+  }, [appBehaviours, navigation, route.params, onSharedDayConfirm]);
+
   /** Open selected day options menu */
   const onDaySelect = (day: Day) => {
     setSelectedDay(day);
     setDeletedDay(null);
     setEditedDay(null);
     Vibration.vibrate(100);
-    dayOptionsRef.current?.open();
+    selectedDayRef.current?.open();
   };
 
   /** Close selected day options menu */
@@ -62,7 +103,7 @@ const HomeScreen = (): ReactElement | null => {
   /** Prepare deletion confirmation dialog (in response to selection menu choice) */
   const onDayDeletePress = () => {
     // NOTE: Ensure previous modal has finished closing before displaying another
-    dayOptionsRef.current?.close(() => {
+    selectedDayRef.current?.close(() => {
       if (!selectedDay) return;
 
       setDeletedDay(selectedDay);
@@ -93,23 +134,23 @@ const HomeScreen = (): ReactElement | null => {
     );
   };
 
-  /** Cleanup day add/edit dialog */
-  const onDayManageCancel = () => {
-    manageDayRef.current?.close();
-    setEditedDay(null);
-  };
-
-  const onDayAdd = (day: DayNew) => {
+  const onDayManageAdd = (day: DayNew) => {
     dispatch(addDay(day));
 
     manageDayRef.current?.close();
     notify(t("screens:dayAddEdit.dayAddSuccess", { title: day.title }));
   };
 
+  /** Cleanup day add/edit dialog */
+  const onDayManageCancel = () => {
+    manageDayRef.current?.close();
+    setEditedDay(null);
+  };
+
   /** Prepare edit dialog (in response to selection menu choice) */
   const onDayEditPress = () => {
     // NOTE: Ensure previous modal has finished closing before displaying another
-    dayOptionsRef.current?.close(() => {
+    selectedDayRef.current?.close(() => {
       if (!selectedDay) return;
 
       setEditedDay(selectedDay);
@@ -119,7 +160,7 @@ const HomeScreen = (): ReactElement | null => {
     });
   };
 
-  const onDayEdit = (day: Day) => {
+  const onDayManageEdit = (day: Day) => {
     dispatch(updateDay(day));
 
     manageDayRef.current?.close();
@@ -127,10 +168,31 @@ const HomeScreen = (): ReactElement | null => {
     notify(t("screens:dayAddEdit.dayEditSuccess", { title: day.title }));
   };
 
-  const onDayMove = (direction: UpDown) => {
-    if (!selectedDay) return;
+  const onDayMove = (day: Day, direction: UpDown) => {
+    dispatch(moveDay({ id: day.id, direction }));
+  };
 
-    dispatch(moveDay({ id: selectedDay.id, direction }));
+  const onDayShare = async (day: Day) => {
+    const url = Linking.createURL("day/shared", {
+      queryParams: {
+        date: day.date,
+        icon: day.icon,
+        id: day.id,
+        repeats: day.repeats ? "true" : "false",
+        title: day.title,
+        version: `${SHARED_DAY_VERSION}`,
+      },
+    });
+
+    const shareMessage = `A day has been shared with you from My Days! Follow this link to add '${day.title}' to the app!\n\n${url}`;
+
+    // NOTE: Must either close selection dialog before sharing or not close it at all
+    selectedDayRef.current?.close(async () => {
+      await Share.share({
+        // Decided to use message over link as it contains more description (and link is iOS-only)
+        message: shareMessage,
+      });
+    });
   };
 
   return (
@@ -154,6 +216,7 @@ const HomeScreen = (): ReactElement | null => {
           onPress={() => navigation.navigate("SettingsRouter")}
         />
       </AppBar>
+
       <View style={[styles.pageHeader, { backgroundColor: colors.primary }]}>
         <Text style={[styles.pageHeaderDay, { color: colors.onPrimary }]} variant="headlineLarge">
           {t("screens:home.todayDay", { date: today })}
@@ -162,6 +225,7 @@ const HomeScreen = (): ReactElement | null => {
           {t("screens:home.todayDate", { date: today })}
         </Text>
       </View>
+
       <View style={styles.pageContent}>
         <DayList days={days} onItemLongPress={onDaySelect} onScroll={onListScroll} />
       </View>
@@ -170,12 +234,19 @@ const HomeScreen = (): ReactElement | null => {
         visible={fabVisible}
         onPress={() => manageDayRef.current?.open()}
       />
+
+      <AddSharedDayDialog
+        day={sharedDay}
+        visible={!!sharedDay}
+        onCancel={() => setSharedDay(null)}
+        onConfirm={() => onSharedDayConfirm(sharedDay!)}
+      />
       <ManageDaySheet
         ref={manageDayRef}
         day={editedDay}
+        onAdd={onDayManageAdd}
         onCancel={onDayManageCancel}
-        onAdd={onDayAdd}
-        onEdit={onDayEdit}
+        onEdit={onDayManageEdit}
       />
       <DeleteDayDialog
         day={deletedDay}
@@ -184,13 +255,14 @@ const HomeScreen = (): ReactElement | null => {
         onConfirm={onDayDeleteConfirm}
       />
       <SelectedDayModal
-        ref={dayOptionsRef}
+        ref={selectedDayRef}
         day={selectedDay}
         dayPosition={selectedDayPosition}
         onClose={onDaySelectCancel}
         onEdit={onDayEditPress}
         onDelete={onDayDeletePress}
         onMove={onDayMove}
+        onShare={onDayShare}
       />
     </Page>
   );