+++ src/typings/day.types.ts
@@ -1,4 +1,6 @@
-import { Optional } from "./app.types";
+import { MaterialCommunityIcons, Optional } from "./app.types";
+
+export type DayUnit = "day" | "week" | "month" | "year";
 
 /** Day shape */
 export interface Day {
@@ -6,7 +8,7 @@ export interface Day {
   createdAt: string;
   /** Day occurrence */
   date: string;
-  icon?: string;
+  icon?: keyof MaterialCommunityIcons;
   /**
    * Whether interval repeats each year
    *
@@ -17,7 +19,7 @@ export interface Day {
   /** Day name/title */
   title: string;
   /** Time unit used for display purposes */
-  unit: "day" | "week" | "month" | "year";
+  unit: DayUnit;
 }
 
 export type DayNew = Optional<Day, "createdAt">;